requirements.txt
opencv-python
numpy
cvzone
Pillow
mediapipe==0.8.11